limpado arquivos desnecessáris e dependencias desnecessarias
